requirements.txt
python-binance>=1.0.30
rich>=13.0.0

